+++ tinper-acs/nc-grid/demo/demolist/Demo2.js
@@ -337,8 +337,16 @@ const tabLists = [{
     name: '详细信息',
     items: [], 
 }];
+
+//获取表格数据时返回的回调
+function getTableRows(data){
+    console.log('保存数据',data);
+}
   
 class Demo2 extends Component {
+    handleClick = () => {
+        this.body.getTableRows();
+    }
     render () {
         const tableHead = [{
             bordered:true,
@@ -357,18 +365,24 @@ class Demo2 extends Component {
             ]
         }]
         return (
-            <CardTable
-            columns={columns}
-            data={data}
-            moduleId="body"
-            tabLists={tabLists}
-            showMore={true}
-            showMax={true}
-            showCheck= {true} //是否开启多选功能
-            hideSwitch= {() => false} //表格
-            tableHead={tableHead}
-            isEdit={true}
-            />
+            <div>
+                <CardTable
+                columns={columns}
+                data={data}
+                moduleId="body"
+                tabLists={tabLists}
+                showMore={true}
+                showMax={true}
+                showCheck= {true} //是否开启多选功能
+                hideSwitch= {() => false} //表格
+                tableHead={tableHead}
+                isEdit={true}
+                onRef={(ref) => { this.body = ref; }} //获取EditTable组件实例
+                getTableRows={getTableRows}
+                />
+                <button onClick={this.handleClick} style={{marginTop:'30px'}}>aaa</button>
+            </div>
+            
         )
     }
 }