+++ tinper-acs/nc-search-panel/CHANGELOG.md
@@ -1,3 +1,33 @@
+<a name="0.0.11"></a>
+## [0.0.11](https://github.com/tinper-bee/nc-search-panel/compare/v0.0.10...v0.0.11) (2019-08-05)
+
+
+### Features
+
+* **nc-search-panel:** 日期区间 acRangepicker特殊处理 ([fbd375e](https://github.com/tinper-bee/nc-search-panel/commit/fbd375e))
+
+
+
+<a name="0.0.10"></a>
+## [0.0.10](https://github.com/tinper-bee/nc-search-panel/compare/v0.0.9...v0.0.10) (2019-08-05)
+
+
+### Features
+
+* **nc-search-panel:** 支持简单查询 ([8c3e9b4](https://github.com/tinper-bee/nc-search-panel/commit/8c3e9b4))
+
+
+
+<a name="0.0.9"></a>
+## [0.0.9](https://github.com/tinper-bee/nc-search-panel/compare/v0.0.8...v0.0.9) (2019-08-05)
+
+
+### Features
+
+* **nc-search-panel:** 按钮对齐方式 ([0e07dc8](https://github.com/tinper-bee/nc-search-panel/commit/0e07dc8))
+
+
+
 <a name="0.0.8"></a>
 ## [0.0.8](https://github.com/tinper-bee/nc-search-panel/compare/v0.0.7...v0.0.8) (2019-08-05)